Iperf: add option to use iperf3
iperf3 is based the original iperf and support similiar command line
and have a similiar output.

This commit allows to pass option to Iperf test module to use iperf3
binary instead of iperf and adds additional error catcher from iperf3.
In iperf probelmatic input is just ignore with a print.

Signed-off-by: Roi Dayan <roid@mellanox.com>